pyproject.toml
[project]
name = "skills-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.4",
    "open-autonomy>=0.19.4",
    "twikit>=2.3.3",
    "web3>=7.9.0",
]

[tool.uv]
prerelease = "allow"
